Demo/backend/requirements.txt
Flask
psycopg2-binary
Flask-Cors

